feat: refactor stocks
