+++ src/app/stocks/stocks.component.html
@@ -1,1 +1,122 @@
-<p>stocks works!</p>
+<div class="p-4">
+    <span class="flex space-x-3">
+        <mat-icon class="mr-2">store</mat-icon>
+        <h3 class="!font-bold">Available Stocks</h3>
+    </span>
+
+    <mat-divider class="!mb-3"></mat-divider>
+
+    <span class="grid grid-cols-2 gap-x-3 gap-y-3">
+        <button
+          class="p-8 w-full border border-gray-500 rounded-lg bg-gray-100 hover:bg-gray-200 hover:border-gray-400"
+        >
+          <span class="font-medium text-3xl">{{ headStock.Petrol | number }}</span>
+          <h3 class="!font-bold !m-0">Current Available Petrol Stock (Ltrs)</h3>
+        </button>
+        <button
+          class="p-8 w-full border border-gray-500 rounded-lg bg-gray-100 hover:bg-gray-200 hover:border-gray-400"
+        >
+          <span class="font-medium text-3xl">{{ headStock.Diesel | number }}</span>
+          <h3 class="!font-bold !m-0">Current Available Diesel Stock (Ltrs)</h3>
+        </button>
+        
+    </span>
+
+    <mat-divider class="!my-3"></mat-divider>
+
+    <span class="flex justify-between">
+        <span class="font-medium">Fuel Stocks summary</span>
+        <button
+          class="border-green-700 border p-2 rounded text-green-700 hover:bg-green-700 hover:text-white"
+          (click)="openAddNewModal(addNewStockModal)"
+        >
+          Add New Stock
+        </button>
+    </span>
+
+    <table
+        mat-table
+        [dataSource]="stocks"
+        class="mat-elevation-z8 w-full mt-3 h-[100px] overflow-y-scroll"
+    >
+        <!--- Note that these columns can be defined in any order.
+            The actual rendered columns are set as a property on the row definition" -->
+
+        <!-- index Column -->
+        <ng-container matColumnDef="index">
+        <th mat-header-cell *matHeaderCellDef>#</th>
+        <td mat-cell *matCellDef="let stock; let i = index">
+            {{ i + 1 }}
+        </td>
+        </ng-container>
+
+        <!-- fuel type Column -->
+        <ng-container matColumnDef="fuelType">
+          <th mat-header-cell *matHeaderCellDef>Fuel Type</th>
+          <td mat-cell *matCellDef="let stock">
+              <span class="font-medium">{{ stock.fuelType }}</span>
+          </td>
+        </ng-container>
+
+        <!-- amount Column -->
+        <ng-container matColumnDef="amount">
+        <th mat-header-cell *matHeaderCellDef>Fuel Amount (Ltrs)</th>
+        <td mat-cell *matCellDef="let stock">
+            {{ stock.fuelAmount | number }} Ltrs
+        </td>
+        </ng-container>
+
+        <!-- date Column -->
+        <ng-container matColumnDef="date">
+        <th mat-header-cell *matHeaderCellDef>Added Date</th>
+        <td mat-cell *matCellDef="let stock">
+            {{ stock.addedDate }}
+        </td>
+        </ng-container>
+
+        <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+        <tr
+            mat-row
+            *matRowDef="let row; columns: displayedColumns"
+            class="cursor-pointer hover:bg-grey-3"
+        ></tr>
+    </table>
+
+<!-- Add new stock - modal -->
+  <ng-template #addNewStockModal>
+    <h2 matDialogTitle>Add New Fuel Stock</h2>
+    <mat-dialog-content>
+      <mat-form-field appearance="outline" class="w-full">
+        <mat-label>Fuel Type *</mat-label>
+        <mat-select [(ngModel)]="fuelType">
+          <mat-option *ngFor="let fuelType of fuelTypes" [value]="fuelType">{{
+            fuelType
+          }}</mat-option>
+        </mat-select>
+      </mat-form-field>
+
+      <mat-form-field appearance="outline" class="w-full">
+          <mat-label>Fuel Amount (Ltrs) *</mat-label>
+          <input matInput placeholder="Fuel Amount" [(ngModel)]="fuelAmount">
+      </mat-form-field>
+    </mat-dialog-content>
+    <mat-dialog-actions align="end">
+      <span class="flex space-x-3">
+        <button
+          matDialogClose
+          class="border-green-700 border p-2 rounded text-green-700 hover:bg-green-700 hover:text-white w-32"
+          (click)="addFuelStock()"
+        >
+          Save
+        </button>
+        <button
+          matDialogClose
+          class="border-red-700 border p-2 rounded text-red-700 hover:bg-red-700 hover:text-white w-32"
+        >
+          Cancel
+        </button>
+      </span>
+    </mat-dialog-actions>
+  </ng-template>
+
+</div>
